+++ protected/controllers/StylesheetColorController.php
@@ -52,25 +52,10 @@ class StylesheetColorController extends Controller
 	}
 
 	/**
-	 * Displays a particular model.
-	 * @param integer $id the ID of the model to be displayed
-	 */
-	public function actionView($id)
-	{
-		if (Yii::app()->authManager->checkAccess('viewStylesheet', Yii::app()->user->id)) {
-			$this->render('view',array(
-				'model'=>$this->loadModel($id),
-			));
-		} else {
-			throw new CHttpException(403,'You are not authorized to perform this action.');
-		}
-	}
-	
-	/**
 	 * Creates a new model.
 	 * If creation is successful, the browser will be redirected to the 'view' page.
 	 */
-	public function actionCreate($ss_id, $ss_code)
+	public function actionCreate($ss_id)
 	{
 		$model=new StylesheetColor;
 		$model->ss_id = $ss_id;
@@ -83,7 +68,7 @@ class StylesheetColorController extends Controller
 		$log_entry->action_type = 'create';
 		$log_entry->user = Yii::app()->user->id;
 		$log_entry->ss_id = $ss_id;
-		StylesheetColorController::insert($model, $cc_model, $log_entry, $ss_code, true);
+		StylesheetColorController::insert($model, $cc_model, $log_entry, true);
 	}
 	
 	/**
@@ -91,7 +76,7 @@ class StylesheetColorController extends Controller
 	 * If update is successful, the browser will be redirected to the 'view' page.
 	 * @param integer $id the ID of the model to be updated
 	 */
-	public function actionUpdate($id, $ss_code)
+	public function actionUpdate($id)
 	{
 		$model=$this->loadModel($id);
 		$cc_model= ColorCode::model()->findByPk($model->color_code);
@@ -100,16 +85,15 @@ class StylesheetColorController extends Controller
 		$log_entry->action_type = 'create';
 		$log_entry->user = Yii::app()->user->id;
 		$log_entry->ss_id = $model->ss_id;
-		StylesheetColorController::insert($model, $cc_model, $log_entry, $ss_code, false);
+		StylesheetColorController::insert($model, $cc_model, $log_entry, false);
 	}
 	
-	public function insert($model, $cc_model, $log_entry, $ss_code, $new)
+	public function insert($model, $cc_model, $log_entry, $new)
 	{
 		$ss_model = Stylesheet::model()->findByPk($model->ss_id);
-		if (Yii::app()->authManager->checkAccess('updateStylesheet', Yii::app()->user->id) ||
-		(Yii::app()->authManager->checkAccess('updateOwnStylesheet', Yii::app()->user->id) && Yii::app()->user->id ==$ss_model->user_id)
-		) {
-			
+		
+		// We're actually modifying the mother sheet, so it's an update
+		if ($this->can('update', $ss_model)) {
 			// Uncomment the following line if AJAX validation is needed
 			$this->performAjaxValidation(array($model, $cc_model));
 	
@@ -154,7 +138,7 @@ class StylesheetColorController extends Controller
 			$this->render(($new)? 'create' : 'update',array(
 				'model'=>$model,
 				'cc_model' => $cc_model,
-				'ss_id' => $model->ss_id, 'ss_code' => $ss_code
+				'ss_model' => $ss_model
 			));
 		} else {
 			throw new CHttpException(403,'You are not authorized to perform this action.');
@@ -170,9 +154,7 @@ class StylesheetColorController extends Controller
 	public function actionDelete($id, $ss_id)
 	{
 		$ss_model = Stylesheet::model()->findByPk($ss_id);
-		if (Yii::app()->authManager->checkAccess('updateStylesheet', Yii::app()->user->id) ||
-		(Yii::app()->authManager->checkAccess('updateOwnStylesheet', Yii::app()->user->id) && Yii::app()->user->id ==$ss_model->user_id)
-		) {
+		if ($this->can('update', $ss_model)) {
 			$this->loadModel($id)->delete();
 	
 			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
@@ -197,7 +179,7 @@ class StylesheetColorController extends Controller
 	 */
 	public function actionIndex($ss_id)
 	{
-		if (Yii::app()->authManager->checkAccess('viewStylesheet', Yii::app()->user->id)) {
+		if ($this->can('view')) {
 			$dataProvider = StylesheetColorController::getStylesheetColors($ss_id);
 			$ss_model = Stylesheet::model()->findByPk($ss_id);
 		
@@ -210,25 +192,6 @@ class StylesheetColorController extends Controller
 		}
 	}
 	
-
-	/**
-	 * Manages all models.
-	 */
-	public function actionAdmin()
-	{
-		if (Yii::app()->authManager->checkAccess('adminStylesheet', Yii::app()->user->id)) {
-			$model=new StylesheetColor('search');
-			$model->unsetAttributes();  // clear any default values
-			if(isset($_GET['StylesheetColor']))
-				$model->attributes=$_GET['StylesheetColor'];
-	
-			$this->render('admin',array(
-				'model'=>$model,
-			));
-		} else {
-			throw new CHttpException(403,'You are not authorized to perform this action.');
-		}
-	}
 
 	/**
 	 * Returns the data model based on the primary key given in the GET variable.
@@ -256,5 +219,13 @@ class StylesheetColorController extends Controller
 			echo CActiveForm::validate($model);
 			Yii::app()->end();
 		}
+	}
+	public function can ($resp, $model=NULL) {
+		return (Yii::app()->authManager->checkAccess("{$resp}Stylesheet", Yii::app()->user->id)||
+		(
+				$model !== NULL &&
+				($resp=='update' || $resp== 'delete') &&
+				Yii::app()->authManager->checkAccess("{$resp}OwnStylesheet", Yii::app()->user->id)&&
+				Yii::app()->user->id == $model->user_id));
 	}
 }
